src/main/java/sprada/solwit/task/service/ReaderService.java
package sprada.solwit.task.service;

import org.springframework.stereotype.Service;

import java.io.*;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.ArrayList;
import java.util.List;

@Service

public class ReaderService {


    public List<String> parseFile(String fileName) {
        List<String> lines = new ArrayList<>();
        try {
            lines = Files.readAllLines(Paths.get(fileName));
        } catch (IOException e) {
            //Could not parse the file
            //Maybe throw an exception
        }
        return lines;
    }

    public InputStream readFile(String path){
        InputStream fis = null;
        try {
            fis = new FileInputStream(path);
        } catch (FileNotFoundException e) {
            e.printStackTrace();
        }
        return fis;
    }


}
